requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.8.0
entrypoints==0.4
Faker==20.0.3
favicon==0.7.0
filelock==3.13.1
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.16.4
ibm-cos-sdk==2.13.0
ibm-cos-sdk-core==2.13.0
ibm-cos-sdk-s3transfer==2.13.0
ibm-watson-machine-learning==1.0.330
idna==3.4
importlib-metadata==6.8.0
InstructorEmbedding==1.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
langchain==0.0.336
langsmith==0.0.64
llama-cpp-python==0.1.9
llama-index==0.9.1
lomond==0.3.3
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.3.0
packaging==23.2
pandas==1.5.3
Pillow==9.5.0
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.4
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.30.0
rich==13.7.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
streamlit==1.24.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-option-menu==0.1.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.18
validators==0.22.0
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
